requirements.txt
boto3==1.34.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0